Modify extractfromlist in Embeddings.jl to return only Ints.
Division in julia appears to automatically return the Float64 type.
Since extractfromlist should always return a valid index the Tuples it
returns must contain only Integer elements.

+++ utils/Embeddings.jl
@@ -43,7 +43,7 @@ function extractfromlist(state::Int, sizes)
             continue
         end
 
-        steps = stepsleft / smallstep
+        steps = Int(stepsleft / smallstep)
         extstate[i] = steps
         counted += stepsleft
